requirements.txt
gtts
pillow
